Moved some postMessage utils to @deephaven/utils (DH-19307)
