+++ packages/jsapi-components/src/useBroadcastChannel.ts
@@ -1,16 +1,12 @@
-import {
-  BROADCAST_CHANNEL_NAME,
-  isMessage,
-  type Message,
-} from '@deephaven/jsapi-utils';
+import { BROADCAST_CHANNEL_NAME } from '@deephaven/jsapi-utils';
 import Log from '@deephaven/log';
-import { EMPTY_FUNCTION } from '@deephaven/utils';
+import { EMPTY_FUNCTION, isMessage, type PostMessage } from '@deephaven/utils';
 import { useEffect, useMemo } from 'react';
 
 const log = Log.module('useBroadcastChannel');
 
 export function useBroadcastChannel(
-  onEvent: (event: MessageEvent<Message<unknown>>) => void = EMPTY_FUNCTION,
+  onEvent: (event: MessageEvent<PostMessage<unknown>>) => void = EMPTY_FUNCTION,
   name = BROADCAST_CHANNEL_NAME
 ): BroadcastChannel {
   const channel = useMemo(() => new BroadcastChannel(name), [name]);
