+++ packages/jsapi-utils/src/SessionUtils.ts
@@ -1,7 +1,8 @@
 import type { dh as DhType } from '@deephaven/jsapi-types';
 import Log from '@deephaven/log';
+import { requestParentResponse } from '@deephaven/utils';
 import { nanoid } from 'nanoid';
-import { requestParentResponse, SESSION_DETAILS_REQUEST } from './MessageUtils';
+import { SESSION_DETAILS_REQUEST } from './MessageUtils';
 import NoConsolesError, { isNoConsolesError } from './NoConsolesError';
 
 const log = Log.module('SessionUtils');
