+++ packages/utils/src/index.ts
@@ -11,6 +11,7 @@ export * from './EventTargetShimUtils';
 export * from './hasRequiredString';
 export * from './isNonEmptyString';
 export * from './isNotNullOrUndefined';
+export * from './MessageUtils';
 export { default as Pending } from './Pending';
 export * from './PromiseUtils';
 export * from './Asserts';
