+++ packages/utils/tsconfig.json
@@ -6,5 +6,5 @@
   },
   "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "src/**/*.jsx"],
   "exclude": ["node_modules", "src/**/*.test.*", "src/**/__mocks__/*"],
-  "references": [{ "path": "../test-utils" }]
+  "references": [{ "path": "../log" }, { "path": "../test-utils" }]
 }
